fixed bug with casting

+++ arduino/commands.h
@@ -97,7 +97,7 @@ void sendbattvoltage(serdata data) {
 }
 
 void setmotorspeed(serdata data) {
-  int32_t new_ltime = data[4] + (data[5] << 8) + ((int32_t) data[6] << 16) + ((int32_t) data[7] << 24);
+  int32_t new_ltime = (uint32_t) data[4] + ((uint32_t) data[5] << 8) + ((uint32_t) data[6] << 16) + ((uint32_t) data[7] << 24);
   int32_t new_rtime = (uint32_t) data[0] + ((uint32_t) data[1] << 8) + ((uint32_t) data[2] << 16) + ((uint32_t) data[3] << 24);
 
   navstate = 2;
